Remove cooler.get() and cooler.read()
Problem: These methods were implemented so that we could use promises in
SSB without the hassle of `util.promisify()` sprinkled everywhere. They
weren't ergonomic or fun, but they were the best we had until promises
were natively supported by SSB-Client (in the MuxRPC module).

Solution: Remove cooler.get() and cooler.read() to use the native
promise support that should be added in the MuxRPC module once
https://github.com/ssbc/muxrpc/pull/58 is merged. This is more ergonomic
and reduces the code we have to write, but *does* need to wait until
that PR is merged.