Update NPM dependencies
Problem: We haven't had a release in a while and our dependencies are
getting old.

Solution: Upgrade them all. I've meticulously read the release notes for
breaking changes and I didn't notice any problems, plus this seems to
run fine, but I'd like to keep this in `master` for a few days before
making a new release.